Remove use of NumberStringUtils class
